+++ UniversityManagementSystem.Domain/Entities/AcademicEntity/Major.cs
@@ -4,9 +4,9 @@ namespace UniversityManagementSystem.Domain.Entities.AcademicEntity
 {
     public class Major : BaseEntity
     {
-        public short MajorId { get; private set; }
+        public ushort MajorId { get; private set; }
         public Guid Guid { get; private set; }
         public string Name { get; set; }
-        public short FacultyId { get; private set; }
+        public sbyte FacultyId { get; private set; }
     }
 }
