+++ UniversityManagementSystem.Domain/Entities/OtherEntity/Nation.cs
@@ -2,7 +2,7 @@
 {
     public class Nation : BaseEntity
     {
-        public short NationId { get; private set; }
+        public sbyte NationId { get; private set; }
         public Guid Guid { get; private set; }
         public string Name { get; set; }
     }
